Correct mpi time functions offset handling.
There are several places in MPI where we permit a user to provide
a custom time offset overriding the timezone. To do this, we
mutilate UTC time before passing it to the `MUCKTIME()` macro.
However, we were doubling the offset of `get_tz_offset()` in
these cases rather than negeting it.

An cleaner solution to this would be using gmtime(), but we don't
know if that's available for our present compatibility goals.

Some historical tickets:
http://sourceforge.net/p/fbmuck/bugs/257/
http://sourceforge.net/p/fbmuck/bugs/271/

Testing done:
Create a MUCK with non-UTC localtime and test mpi codes before and
after the change:

   {ftime:%m-%d %X,2}
   {date:0}
   {ftime:%Y-%m-%d %X,0,0}  # Should be epoch.
   etc.