Make tunable defaults changeable after first run
Add syntax for default tunable values:
> Database (custom, then default)
ssl_min_protocol_version=TLSv1.2
%ssl_min_protocol_version=None
> When using @tune
@tune ssl_min_protocol_version=TLSv1.2
@tune %ssl_min_protocol_version

This works in the database and in prim_setsysparm.  When setting a
tunable to default by prefixing with '%', the given value is ignored.
Default values are recorded to the database, but ignored on loading.
This allows for compiled defaults to change after install, e.g. for
tightening SSL security, while letting the sys-admin easily find and
inspect new defaults, overriding them if desired.

Uncommenting defaults to modify them doesn't require adjusting the
fixed-length parameter count in the database header.  New defaults
are also easy to spot.

Add support for tracking 'isdefault' and a default value to all types
of tune_*_entry.  Specify defaults in the tune lists.  This is needed
to support resetting a tunable to default without restarting Fuzzball.