Add help for SYSPARM param consistent_lock_source
Add "consistent_lock_source" to mufman.raw and related generated
help files.

In the future, this portion of the help should be automatically
generated, but for now it may be best to keep it in sync.  Otherwise,
to learn about it one must read the source or have a running server.