Re-add 'make help', remove HTML on 'make clean'
Uncomment the 'make help' functionality in Makefile.in, using a new
HTML_DOCS_DIR variable, currently set to 'docs_html'.  This makes the
generated help functionality more obvious to new contributors and
hopefully reduces confusion.

In the future, this should be cleaned up so all generated docs are
made as part of 'make all', and help.txt/man.txt should be removed
from the repository.

Remove generated docs as part of 'make clean', ignore the generated
docs folder in Git.