Fix a rare color issue in advancedb WHO
I noticed, on Here Lie Monsters, that when a player has been connected for over 100 days, and their idle time is over 60 seconds but under 600 seconds (usually uncolored!) that the rainbow text applied to their 'on for' time can get applied to their idle time.  This diff makes it so that the idle time is always uncolored for that time period, regardless of other factors.
Diff partially tested.  I didn't leave a MUCK running for 100 days, nor enabled the rainbow text at an earlier time.  Diff works correctly though.