Fix time_t to int conversion
Spotted this in appveyor's output, after ipv6 was enabled for all builds.  Also checked the ipv4 version, and fixed it too.
While there, fixed some stylistic discrepancies between the two, and made a comment more readable.  I wonder if the situation described in the comment still exists as of this decade or not, but I won't touch that.