Switch to INT_MAX (#291)
* Update p_stack.c

* Update p_math.c

* Update mfuns.c

* Update compile.c

* Update fbmath.c

* Get rid of MAXINT and MININT

Use standard INT_MAX and INT_MIN

* Fix documentation for RANDOM

Turns out the C function random() has nothing to do with INT_MAX or MAXINT, but has a set ceiling.  random() is also not standard C, but a common Unix function from BSD.
In fact, we have ifdefs that mean this could be based on either random() or rand() depending on whether the system has random() or not!  So it's impossible to have correct documentation here until those ifdefs are resolved.

* Update man.txt

* Update mufman.html