Fix page's documentation
The example for using page with a message is wrong. Using minimaldb/fuzzball's page, rather than a MUF page, shows this output on the recipient's side:
One pages from Room Zero: "uwu"

I'd move examples in help pages to their own example section at the bottom (like in man pages) or maybe even remove them. But I retained them.
On the topic of man pages, I think if there's a way to switch to them, that'd be awesome. The current documentation is incredibly inconsistent, and while consistency isn't a goal on its own, the level of inconsistency here sometimes interferes with readability.