Fix size_t to int comparison in prochelp
bufsize is always a size_t, and pos is only ever compared with bufsize.  Up pos from int to size_t (which CI checks say is an unsigned long)
I still think strcpyn could be replaced with some usage of strncpy...  But this is just a small easy fix for this comparison warning for now.