Fix a warning C4018 (#322)
* Fix a warning C4018

https://github.com/fuzzball-muck/fuzzball/blob/master/src/look.c#L1084
https://ci.appveyor.com/project/fuzzball-muck/fuzzball/builds/22582572#L265
This will likely introduce new warnings, which I'll have to check via CI since I don't have access to Windows.  I also fixed minor formatting/alignment. :P

* Fix formatting

Uses tabs instead of spaces for the indentation, which are far easier to indent with.
The file also uses tabs for aligning the comments though...  That I consider more dubious, BUT, it was already like that and that's not my change.
I also switched the comment after the endif to only be separated by a single space, as per the only styleguide I read which had anything to say about it.
https://www.freebsd.org/cgi/man.cgi?query=style&apropos=0&sektion=9&manpath=FreeBSD+12.0-RELEASE+and+Ports&arch=default&format=html