Move documentation to docs folder
The intention of this diff is to move all the documentation that humans (and fluffy animals) might want to read to a single top-level folder in the tarball, so they're easy to find.  Then, only files that fuzzball will read are copied back into game/data
For example, none of the .raw or .html files are read by fuzzball.  Thus, they don't get put into game/data
Since the .raw don't get installed to INSTALL_HELPDIR later on (implying they aren't meant to be read by documentation-seeking humans) but also aren't read by fuzzball, I treat them as source files and moved them to src

Unrelatedly, use autoconf's docdir folder instead of building the path ourselves off of datadir, for INSTALL_HELPDIR.  The folder remains the same.  This should improve things for as long as we use autoconf, in terms of user customization.