Include standard C headers (#335)
* Always include stdint.h

It's part of C99 (and inherently part of posix).  We use C99.  Get rid of some unfortunate typedefs while we're at it.
My intention is to get rid of some HAVE_* defines for headers part of the C standard.

* remove typedefs in fbmath.c

* Remove HAVE_STDLIB_H

* Get rid of HAVE_STRING[S]_H

string.h is a part of C.  HAVE_STRINGS_H we also don't use anymore