Regenerate minimaldb (#345)
* Regenerate minimal.db

I manually deleted all but one sysparm in minimaldb, changed 158 to 1, and fuzzball was able to start.  I then started, did at-shutdown, and voila, minimaldb was made again.
I want to figure out how to make the smallest minimaldb possible, but UNTIL THEN, this at least makes all sysparms show up as [default] when they are default.  That is a good success.

* update minimaldb again

We have two of the same file

* Removed props.

* Removed props.