Delete game/data/minimal.db (#346)
We already have the same file in the dbs/ folder.
It also turns out the documentation is already there for this, even if it uses basedb as an example, it works with minimaldb too. :P  Oops!  Shows how little I test with dbs other than minimal.
This is the entire PR.