add changesfb7 and other documentation (#377)
* Add changesfb7

This is useful info.  I have a TODO embedded in it right now, which can be removed later.
It has a beta 2 section to be filled out before beta 2 is released.

* Update Makefile.in

include just changesfb7 in the info folder
In the future,  changesfb6  (and changesfb5) MIGHT be added back there.  But importantly, changesfb7 should be there.  changesfb4 and earlier don't need to be, at all.

* Update appveyor.yml

:D

* create historical folder

* Add old changes files

Sourced from the fbmuck-6.17 tarball.  Those are actually more up-to-date than the changes files removed from the repository in this commit:
https://github.com/fuzzball-muck/fuzzball/commit/afe8c478e8c5f12031e4c915b1b9529d8445f27e

* Delete deletethis

* Rename CHANGESfb to changesfb

This makes them easier to copy into the info folder, and has no downside AFAIK

* Rename CHANGESfb2 to changesfb2

* Rename CHANGESfb3 to changesfb3

* Rename CHANGESfb4 to changesfb4

* Rename CHANGESfb5 to changesfb5

* Rename CHANGESfb6 to changesfb6

* Update index.html

* Create oldlicense

This provides a history of Fuzzball's licensing, and provides some old licensing info that should exist in this repository.  Files sourced from fb5.68

* Fix case of first person in changesfb7 todo