switch from random() to arc4random()
OSes without arc4random() will use rand()
The usage of arc4random_uniform() with that argument is to guarantee the results have the same range as rand().
arc4random() has no seed function, so we just call srand(), which does not affect arc4random()'s results.  Hey, won't this affect behavior?  The answer is actually no!  SRANDOM() is only ever called once, with a seed of getpid(), meaning repeatability wasn't even desired there.  In the code, only actual srand() is used when a deterministic PRNG is desired.
I've found both rand() and RANDOM() in the code, but not random().  That's a portable mix, already.