Remove colorisation of WIZ status (#391)
* Remove colorisation of WIZ status

This is a no-no.  The status prop isn't protected in any way at all, so anyone can set themselves to have the WIZ status.  That's fine, but it shouldn't be given its own color in ws.
This should make impersonating a wiz by accident or on purpose less likely.

* Remove highlighting of WIZ status

Remove related prop.  Fix status's documentation.