Create fbmuck.h
Moving the dbref typedef from config.h to a new file, fbmuck.h, where nonconfigurable but program-wide stuff can reside.
At least, I don't think this typedef is meant to be configurable.  The MUCK wouldn't build if it was changed, unfortunately. :P

The define is intentionally FBMUCK_H and not _FBMUCK_H
Having defines start with an underscore isn't a good idea.  I believe they're reserved for C implementations, and I can double-check on that if necessary.
TODO: change our current defines to not start with an _ underscore.