Fix our defined macros (#394)
* fix our defined macros

remove the _ underscore as words beginning with that and a capital letter after are supposed to be reserved to the implementation

* Update boolexp.h

* Update color.h

* Update commands.h

* Update compile.h

* Update config.h

* Update crt_malloc.h

* Update db.h

* Update dbsearch.h

* Update debugger.h

* Update defaults.h

isn't a lot of this configurable?  Should this some day be merged with config.h?  I think so.  But config.h should be shrunk first.
unless... i think some of these may be tunes, aren't they

* Update diskprop.h

* Update edit.h

* Update events.h

* Update fbmath.h

* Update fbsignal.h

* Update fbstrings.h

* Update fbtime.h

* Update game.h

* Update hashtab.h

* Update inst.h

* Update interface.h

* Update interface_ssl.h

Woah.  I think ```//``` are perfectly fine for genuinely single-line comments.  I left them where they're for single-line comments here, and replaced them with normal C comments for multi-line ones.
I also removed some indentation from  define SSL_UNSUPPORTED_PROTOCOL which I believe is incorrect given all the ifdef afterwards not being so indented.

* Update interp.h

* Update log.h

* Update match.h

* Update mcp.h

* Update mcpgui.h

* Update mcppkg.h

* Update mfun.h

* Update move.h

* Update mpi.h

* Update msgparse.h

* Update mufevent.h

* Update p_array.h

* Update p_connects.h

* Update p_db.h

* Update p_error.h

* Update p_float.h

* Update p_math.h

* Update p_mcp.h

* Update p_misc.h

* Update p_props.h

* Update p_regex.h

* Update p_stack.h

* Update p_strings.h

* Update player.h

* Update predicates.h

* Update props.h

* Update sha1.h

* Update timequeue.h

* Update tune.h

* Update tunelist.h

* Update diskprop.h

* Update fbmath.h