Remove TIME_MINUTE/HOUR/DAY macros (#416)
* Don't use TIME_MINUTE/HOUR/DAY macros

They're only used in this file.  I personally think the readability required is done better via good commenting than via macros.  It also reminds users that the final value genuinely is in seconds, and they can use any number of seconds they want.
We also save a few steps of multiplication, though that's not the impetus here.

If you think this is a bad idea, let me know!

I also chose to align everything. (with spaces, of course)
I also added a blank line to emphasise that PNAME_HISTORY_REPORTING is a boolean, not a time.

* TIME_MINUTE elimination day

Remove some now-unused macros