Recommitted advancedb.db as an extract from an SQLite database.
The main difference is the type of string properties. Sometimes it was "2", sometimes it was "10", and I have yet to figure out why. I have standardized on "2" for now, but we can change it to whatever is appropriate.

Having the old COMPRESSED flag in the mix (https://fbmuck-devel.livejournal.com/8495.html) would explain it, but we don't have that anymore. Right?

I must be missing something.