ci: Move Windows to VS 2019, add rendered README to build artifact (#578)
* docs: Inline Markdown URLs for readability

Inline URLs within the Markdown files for greater readibility when
viewing as plain text.  This means some URLs are duplicated, but a
find-and-replace shouldn't be too difficult in the future.

Clean up the status badges on the main README file.

* ci: Remove Unix file from Windows build artifact

Remove "restart.in" from Windows build artifact; it's not needed.

* ci: Migrate Windows compiler to Visual Studio 2019

Move to the Visual Studio 2019 image from AppVeyor.  Fuzzball seems to
function as expected and is used by the Windows maintainer (Keet Fox).

This is needed to access PowerShell 7.0 on AppVeyor Windows CI
build environment images.

* ci: Add Windows README as text, HTML, add source

Include Windows README files as a plain-text file converted to Windows
line endings, and render the Markdown to HTML files as well, modifying
".md" links to ".md.htm" instead.

Due to AppVeyor not using PowerShell 7.0 by default, include a
workaround that calls the non-default PowerShell install if system
PowerShell version isn't high enough.

Move "Running" section to top now that README is included.