ONLINE and ONLINE_ARRAY ignore duplicates
Only one dbref per player will be returned, even if there are many descriptors
connected to it.

This still does not match the old behavior, where players would be returned
in order of connection, more or less.  Instead, the list is ordered by dbref.
This could be rolled back fairly easily, but the options are:

  1. Populate the lookup dict twice, first to count how large the output array
     needs to be, and then second to fill it after creating it.
  2. Use array_appenditem, which spams realloc() because it only adds one
     additional element each time.

Both of these feel kind of bad. Additionally,

  1. The original behavior was never documented, and
  2. The only program I know of that ONLY displayed connected players, rather
     than multiple connections 'WHO'-style, sorts the list itself anyway.