Fix process_command skipping null terminator
For @co* and @re*, the command parser gets ahead of itself and skips the
null terminator. This is relatively benign, since it never accesses
anything close to the end of the command buffer. The only effect seems
to be that it can incorrectly match the previous command using an
invalid prefix:

    > @recycle
    I don't understand ''.
    > @r
    I don't understand ''.
    > @readlock
    Read Lock: *UNLOCKED*
    > @r
    Read Lock: *UNLOCKED*