Fix some compiler warnings
This only fixes some minor warnings. It leaves most of the very spammy
gcc 'format truncation' errors, which come from two main sources:

(1) the resolver, which seems to play very fast-and-loose with buffer
    lengths, and

(2) any time we iterate over a propdir, where we concatenate the base
    dir and the subkey.

The former can be fixed, but the latter is almost impossible to eliminate
because we can't know the basedir or subkey lengths, just that their sum
plus delimiter has to fit in a single buffer.